+++ src/main/java/me/jinjjahalgae/domain/user/usecase/delete_my_account/DeleteMyAccountUseCase.java
@@ -1,4 +1,4 @@
-package me.jinjjahalgae.domain.user.usecase.interfaces;
+package me.jinjjahalgae.domain.user.usecase.delete_my_account;
 
 public interface DeleteMyAccountUseCase {
     void execute(Long userId);
