+++ src/main/java/me/jinjjahalgae/domain/user/usecase/get_my_info/GetMyInfoUseCase.java
@@ -1,0 +1,7 @@
+package me.jinjjahalgae.domain.user.usecase.get_my_info;
+
+import me.jinjjahalgae.domain.user.usecase.common.dto.MyInfoResponse;
+
+public interface GetMyInfoUseCase {
+    MyInfoResponse execute(Long userId);
+}
