--- src/main/java/me/jinjjahalgae/domain/user/usecase/interfaces/GetMyInfoUseCase.java
@@ -1,7 +1,0 @@
-package me.jinjjahalgae.domain.user.usecase.interfaces;
-
-import me.jinjjahalgae.domain.user.dto.MyInfoResponse;
-
-public interface GetMyInfoUseCase {
-    MyInfoResponse execute(Long userId);
-}
