--- src/main/java/me/jinjjahalgae/domain/user/usecase/interfaces/UpdateMyInfoUseCase.java
@@ -1,8 +1,0 @@
-package me.jinjjahalgae.domain.user.usecase.interfaces;
-
-import me.jinjjahalgae.domain.user.dto.MyInfoResponse;
-import me.jinjjahalgae.domain.user.dto.UpdateMyInfoRequest;
-
-public interface UpdateMyInfoUseCase {
-    MyInfoResponse execute(Long userId, UpdateMyInfoRequest request);
-}
