+++ src/main/java/me/jinjjahalgae/presentation/api/UserController.java
@@ -2,11 +2,11 @@ package me.jinjjahalgae.presentation.api;
 
 import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
-import me.jinjjahalgae.domain.user.dto.MyInfoResponse;
-import me.jinjjahalgae.domain.user.dto.UpdateMyInfoRequest;
-import me.jinjjahalgae.domain.user.usecase.interfaces.DeleteMyAccountUseCase;
-import me.jinjjahalgae.domain.user.usecase.interfaces.GetMyInfoUseCase;
-import me.jinjjahalgae.domain.user.usecase.interfaces.UpdateMyInfoUseCase;
+import me.jinjjahalgae.domain.user.usecase.common.dto.MyInfoResponse;
+import me.jinjjahalgae.domain.user.usecase.update_my_info.dto.UpdateMyInfoRequest;
+import me.jinjjahalgae.domain.user.usecase.delete_my_account.DeleteMyAccountUseCase;
+import me.jinjjahalgae.domain.user.usecase.get_my_info.GetMyInfoUseCase;
+import me.jinjjahalgae.domain.user.usecase.update_my_info.UpdateMyInfoUseCase;
 import me.jinjjahalgae.global.common.CommonResponse;
 import me.jinjjahalgae.global.security.jwt.CustomJwtPrincipal;
 import me.jinjjahalgae.presentation.api.docs.user.UserControllerDocs;
